test(library): harden Prompt evidence driver
