+++ Docs/superpowers/reviews/evidence/task-22033/README.md
@@ -1,10 +1,10 @@
 # TASK-22033 live verification evidence
 
 The production-CSS Textual harness completed successfully against the real
-Library screen and isolated Prompt databases on 2026-08-25, including a fresh
-post-rebase run. The driver imported the package from the
+Library screen and isolated Prompt databases on 2026-08-26, including a fresh
+post-review run. The driver imported the package from the
 `library-prompts-reader` worktree and used the final scratch profile rooted at
-`/private/tmp/task22033-live-rebased-20260825`.
+`/private/tmp/task22033-qodo-full.vUtAA0`.
 
 ## Verified journeys
 
@@ -15,8 +15,9 @@ post-rebase run. The driver imported the package from the
 - `80x24`: Library and Prompts collapse, Work expands to `70` columns, and both
   five-column restore grips remain focusable and correctly named.
 - Basic is the default editor mode at every geometry.
-- A Basic-mode edit preserved hidden structured Prompt block metadata, saved a
-  new version, and appeared in Info/history with Local provenance.
+- A Basic-mode edit preserved the complete structured Prompt definition,
+  including every lane, block, and metadata field, saved a new version, and
+  appeared in Info/history with Local provenance.
 - Blank-name validation retained the dirty draft and focused the owning Name
   input with actionable copy.
 - Bulk mode kept the loaded Prompt visible as a labelled, disabled read-only
@@ -31,7 +32,9 @@ post-rebase run. The driver imported the package from the
 
 `summary.json` is the machine-readable ledger. Each journey also has a `.json`
 fact file, a plain rendered `.txt` capture, and a production-CSS `.svg`
-screenshot. `task22033_live_matrix.py` is the reproducible isolated driver.
+screenshot. `task22033_live_matrix.py` is the isolation and selector bootstrap;
+`task22033_live_matrix_runner.py` owns the mounted journeys and deterministic
+database cleanup.
 
 ## Reproduction
 
@@ -39,9 +42,10 @@ Run the driver from the worktree with a fresh `TASK22033_SCRATCH_ROOT`, XDG
 directories, `TLDW_CONFIG_PATH`, and `TASK22033_DATA_DIR`, plus
 `TLDW_TEST_MODE=1`, `TLDW_DISABLE_MODEL_CATALOG_NETWORK=1`, and the worktree on
 `PYTHONPATH`. The driver creates a private scratch config whose `paths.data_dir`
-also stays under that root; it refuses to run when isolation variables are
-missing or when config/data escape the scratch root. It does not replace
-`HOME`.
+also stays under that root, replacing any pre-existing scratch config before
+application imports. It rejects malformed or unsupported journey selectors and
+refuses to run when isolation variables are missing or when config/data escape
+the scratch root. It does not replace `HOME`.
 
 The harness deliberately stubs unrelated Library source services. Their
 unavailable study-backend debug traces do not affect the Prompt journeys and are
