+++ Docs/superpowers/reviews/evidence/task-22033/prompts-100x30.txt
@@ -27,4 +27,4 @@
 
                                                       Use in Console
                                                       More actions
-  / focus search | F6 next pane | esc back to list | F1 help · Ctrl+P palette · Ctrl+Q quit
+  / focus search | F6 next pane | esc focus rail | F1 help · Ctrl+P palette · Ctrl+Q quit
