+++ Docs/superpowers/reviews/evidence/task-22033/prompts-import-work-pane.json
@@ -1,8 +1,8 @@
 {
   "editor_hidden": true,
   "identities": [
-    5315011904,
-    4781808416
+    5110511184,
+    5112084336
   ],
   "import_path_visible": true,
   "items_retained": true,
