+++ Docs/superpowers/reviews/evidence/task-22033/prompts-preservation-history.json
@@ -3,7 +3,7 @@
   "history_title": "Retained history (3)",
   "mode": "info",
   "persisted_version": 3,
-  "prompt_uuid": "34d3f1d2-dd18-4602-91e4-89f93f162846",
+  "prompt_uuid": "f646a68a-cc2e-4cd6-ae07-5d3389e42270",
   "provenance": "Persisted source: Local \u00b7 Prompt \u00b7 structured format",
   "svg": "prompts-preservation-history.svg"
 }
