+++ Docs/superpowers/reviews/evidence/task-22033/prompts-preservation-history.txt
@@ -32,19 +32,19 @@
 │ │                          │          ▊                                                                                                                   ▎  │
 │ │    Quizzes               │          ▊                                                                                                                   ▎  │
 │ │                          │          ▊                                                                                                                   ▎  │
-│ │ ──────────────────────── │          ▊                             v3 · change 7 · 2026-08-26T06:58:34.526000+00:00 · Prompt                             ▎  │
+│ │ ──────────────────────── │          ▊                             v3 · change 7 · 2026-08-27T01:53:05.061000+00:00 · Prompt                             ▎  │
 │ │  Import / Export    ▾    │          ▊                                             Author, System, Definition                                            ▎  │
 │ │    Import…               │          ▊                                                                                                                   ▎▅ │
 │ │    Export                │          ▊                                                                                                                   ▎  │
-│ │ ──────────────────────── │          ▊                             v2 · change 6 · 2026-08-26T06:58:32.508000+00:00 · Prompt                             ▎  │
+│ │ ──────────────────────── │          ▊                             v2 · change 6 · 2026-08-27T01:53:03.174000+00:00 · Prompt                             ▎  │
 │ │  Details            ▸    │          ▊                                                    Description                                                    ▎  │
 │ │                          │          ▊                                                                                                                   ▎  │
 │ │                          │          ▊                                                                                                                   ▎  │
-│ │                          │          ▊                             v1 · change 1 · 2026-08-26T06:58:32.467000+00:00 · Prompt                             ▎  │
+│ │                          │          ▊                             v1 · change 1 · 2026-08-27T01:53:03.132000+00:00 · Prompt                             ▎  │
 │ │                          │          ▊                                                      Created                                                      ▎  │
 │ │                          │          ▊                                                                                                                   ▎  │
 │ │                          │                                                                                                                                 │
 │ │                          │                                                              Use in Console                                                     │
 │ └──────────────────────────┘                                                               More actions                                                      │
 └──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┘
-  / focus search | F6 next pane | esc focus rail | F1 help · Ctrl+P palette · Ctrl+Q quit
+  / focus search | F6 next pane | esc back to list | F1 help · Ctrl+P palette · Ctrl+Q quit
