+++ backlog/tasks/task-22033 - Migrate-Library-Prompts-to-the-adaptive-reader-shell.md
@@ -57,4 +57,6 @@ Post-Qodo hardening: removed the unused test-only PromptReaderState module so Li
 CI derived-artifact review: inspected the exact persistent-diagnostic statement delta in tldw_chatbook/UI/Screens/library_screen.py (two new constant-message debug calls and one shortened existing info call retaining only the internal Prompt integer ID). Confirmed no new sink topology and no interpolation of user content, secrets, paths, or URLs; regenerated Docs/security/production-diagnostic-inventory.json and verified it reproduces at 538 owners, 1,241 TASK-492 calls, 7,359 TASK-494 calls, and 8 sink files.
 
 Post-rebase verification: rebased onto origin/dev at c23113e2e00142d918bd64b56a53f755b07424ce and regenerated the production diagnostic inventory to resolve the generated-artifact conflict from source (537 owners, 1,244 TASK-492 calls, 7,362 TASK-494 calls, 8 sink files). Hardened two order-sensitive mounted-canvas tests to await actual widget/modal/notification settlement instead of fixed pauses. Verified the complete 341-test Prompt canvas file in exact order, with the remaining 314 targeted Prompt tests green in the immediately preceding run; also passed Ruff, compileall, diff checks, and diagnostic inventory reproduction. No new ADR required; this is verification hardening within ADR-086.
+
+Final Qodo evidence hardening: split the isolated bootstrap from the mounted journey runner so configuration and path containment complete before third-party/application imports; validated every CLI journey selector through the shared input-validation boundary and a closed allow-list; always replaced scratch configs with a contained app-data path; deterministically closed every real Prompt database after the Textual host and on failures; documented every public journey; and strengthened Basic-mode preservation evidence to compare the complete structured definition rather than selected user-block fields. Added focused regressions for escaped XDG paths, malformed/unsupported selectors, hostile existing configs, and failure-path database cleanup, then regenerated the full production-CSS evidence matrix. No new ADR required; these changes harden the existing verification harness within ADR-086.
 <!-- SECTION:NOTES:END -->
